+++ src/service/graphql/schema/schemaDirectives/index.ts
@@ -3,6 +3,8 @@ import { SmartCacheDirective } from './SmartCacheDirective'
 import { Translatable } from './Translatable'
 import { TranslatableV2 } from './TranslatableV2'
 
+export { parseTranslatableStringV2, formatTranslatableStringV2 } from './TranslatableV2'
+
 export const nativeSchemaDirectives = {
   auth: Auth,
   smartcache: SmartCacheDirective,